requirements.txt
beautifulsoup4==4.11.1
discord.py==2.3.2
python-dotenv==1.0.1
yt-dlp==2023.12.30